Factor out create_object().
