+++ components/tools/OmeroPy/test/integration/clitest/cli.py
@@ -20,8 +20,10 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 
+import omero
 from omero.cli import CLI
 from omero.plugins.sessions import SessionsControl
+from omero.rtypes import rstring
 
 from library import ITest
 from omero_ext.mox import Mox
@@ -47,6 +49,27 @@ class CLITest(AbstractCLITest):
 
     def setup_method(self, method):
         self.args = self.login_args()
+
+    def create_object(self, object_type):
+        # create object
+        if object_type == 'Dataset':
+            new_object = omero.model.DatasetI()
+        elif object_type == 'Project':
+            new_object = omero.model.ProjectI()
+        elif object_type == 'Plate':
+            new_object = omero.model.PlateI()
+        elif object_type == 'Screen':
+            new_object = omero.model.ScreenI()
+        elif object_type == 'Image':
+            new_object = self.new_image()
+        new_object.name = rstring("")
+        new_object = self.update.saveAndReturnObject(new_object)
+
+        # check object has been created
+        found_object = self.query.get(object_type, new_object.id.val)
+        assert found_object.id.val == new_object.id.val
+
+        return new_object.id.val
 
 
 class RootCLITest(AbstractCLITest):
